refactor(client): modularize IPC driver and mux handling
Splitted monolithic `lib.rs` into distinct modules for `IpcDriver` and `mux`. Centralized multiplexer functionality in `mux.rs` and driver logic in `driver.rs` to improve code organization, readability, and maintainability.
